+++ module-video-details/src/main/java/com/pp/module_video_details/ui/VideoDetailsActivity.kt
@@ -29,7 +29,7 @@ class VideoDetailsActivity :
 
     @JvmField
     @Autowired(name = "resourceId")
-    var resourceId: Int? = 0
+    var resourceId: Long? = 0
 
     @JvmField
     @Autowired(name = "resourceType")
